Fix defining bindings in OR patterns
When type-checking OR patterns we use a "Pattern" structure for each
branch of the OR, tracking variables defined in that branch. This commit
ensures we also propagate the defined variable IDs up to the parent
"Pattern" structure, ensuring MIR has access to these IDs. Failing to do
so would lead to a compiler panic.

This fixes https://gitlab.com/inko-lang/inko/-/issues/260.

Changelog: fixed