Refactor how processes are terminated
Process termination no longer considers the atomic reference count of a
process, instead relying on a "terminate" argument for the
"ProcessFinishTash" instruction.

Prior to this commit, terminating a process could result in a race
condition where the receiving process observed its reference count to be
zero and drop itself, with a parent process scheduling the destructor
for the receiver a moment later, resulting in the VM running a
deallocated process.

This fixes https://gitlab.com/inko-lang/inko/-/issues/276.

Changelog: fixed