Remove alive tracking from the timeout worker
This removes the "alive" flag from the timeout worker, instead relying
on dropping channels to signal the shutting down of this worker. This is
done in preparation for https://gitlab.com/inko-lang/inko/-/issues/275,
as it removes the need for explicitly shutting down a timeout worker,
while also simplifying the internals.

Changelog: changed