Remove alive tracking from the network poller
This is done in preparation for
https://gitlab.com/inko-lang/inko/-/issues/274, as it removes the need
for explicitly shutting down network pollers. Once upon a time embedding
Inko was something I was toying with, and in such an environment it
makes sense to explicitly and cleanly shut down threads. But in light of
wanting to turn Inko into a compiled language, this isn't relevant
anymore, and we can simplify things by just aborting.

Changelog: changed