Disallow processes sending messages to themselves
Processes sending messages to themselves can result in deadlocks,
whether the message's result is awaited immediately or not. Instead of
trying to go through all sorts of hoops to detect this as late as
possible, we'll just check for it when sending messages.

This fixes https://gitlab.com/inko-lang/inko/-/issues/259.

Changelog: fixed