Allow setting a shuffle seed for unit tests
Using a fixed seed makes it easier to reproduce test failures that are
the result of ordering of tests. The seed itself only controls the
initial order tests are sorted in before they're scheduled, it doesn't
control the order in which they finish. If tests must both start and
finish in the exact same order, the `concurrency` option must be set to
`1`, at the cost of increasing test execution times.

Changelog: added