Suspend after exiting a blocking call
When the new scheduler was introduced, processes were allowed to
continue running when exiting a blocking call, and the OS thread
wouldn't transition into a backup thread until the next time it would
pick up work. This could lead to issues such as:

- One OS thread waking up multiple backup threads, if it kept entering
  blocking IO calls.
- Multiple threads using the same queue, while slow threads are
  constantly pushing work back onto the global queue.
- Threads overwriting each others' shared epoch, resulting in some being
  allowed to run for longer or being immediately marked as being too
  slow.
- All threads hanging while work still remains, because they all
  transitioned to backup threads but didn't find any blocked threads.

To fix this, when returning from a builtin function call (currently the
only place where we can perform blocking IO) we check the backup flag.
If set to true, the current process is pushed back onto the global queue
and the OS thread transitions into a backup thread.

See https://gitlab.com/inko-lang/inko/-/issues/278 for more information.

Changelog: fixed