Remove the dirs-next dependency
This crate hasn't been maintained in a while, and we only used it for
obtaining the home directory, which is easy enough to implement
ourselves.

On Windows we expect the USERPROFILE variable to be set, and we don't
have a fallback to SHGetFolderPathW(). As far as I can tell USERPROFILE
is usually set, so it should be good enough. In all cases if the home
directory is explicitly set to an empty string we ignore it, because
there's no scenario in which one would treat it any different from a
home directory not being set at all.

Changelog: changed