Remove glob imports from the language
Blindly importing symbols was something I never really cared for, but
added because "why not". I'm now of the opinion it's better to be
explicit about what you're importing, instead of blindly importing
everything a module has to offer. With that in mind, support for glob
imports is removed in this commit.

Changelog: changed