Support mod.inko and relative imports
When importing a path that leads to a directory, if the directory
contains a mod.inko file this file is imported instead of producing an
error. This allows you to write something like this:

    import 'gitlab.com/foo/json'

The compiler then tries to import the file gitlab.com/foo/json/mod.inko.

We also add support for imports relative to the current module's
directory. This way you can write something like this:

    import 'foo'
    import 'bar'

Regardless of the dependencies added, if the files `foo.inko` and
`bar.inko` exist in the same directory as that of the module importing
them, the import always pulls in those local files. This way you can
build a project without having to store it in a particular path (e.g.
GOPATH for Go) and use imports relative to that path.

Changelog: added