Add support for socket deadlines/timeouts
Socket operations can now be bound by a deadline, resulting in them
timing out once the deadline is crossed.

We use deadlines instead of timeouts as deadlines compose better, and
make it easier to apply a time limit to a group of operations.

Deadlines are set using Socket.timeout_after= and
UnixSocket.timeout_after=. These methods take either an Instant or a
Duration. If a Duration is given it's added to the result of
`Instant.new`. The return value is a guard which clears the deadline
when dropped. Overwriting a deadline results in a panic, as supporting
nested deadlines leads to all sorts of hairy issues (e.g. inner
deadlines being greater than outer deadlines).

Internally our deadlines are converted to the std::time::Instant type of
the Rust standard library. This requires a bit of work as we can't just
fabricate an Instant from an existing integer.

When a deadline is used, the socket is registered with a network poller
as usual, and the process is registered with the timeout thread. Both
the timeout thread and network poller take care to only reschedule a
process they have the appropriate rescheduling rights for. When a
process is rescheduled we check if a timeout expired. If this was the
case we de-register the socket with the network poller and throw an IO
error. By de-registering the socket we ensure the network poller stops
polling it for completion.

Deadlines are only supported for sockets. On Unix systems file deadlines
could be implemented using signals, but this likely results in a messy
setup. On Windows and other platforms it's not clear what approach to
take. The closest cross-platform solution is to close a file after a
timeout, but this prevents using the file from that point on, making it
rather useless.

This fixes https://gitlab.com/inko-lang/inko/-/issues/228.

Changelog: added