Handle negative Instants and add Instant.remaining
Instant.+ and Instant.- can produce negative values depending on the
Duration passed as an argument. As a negative Instant makes no sense
(because it's a monotonic clock starting at some zero point), we now
panic if these methods would produce a negative Instant.

In addition, we add Instant.remaining to calculate the remaining time
until reaching a given Instant. This is useful when an Instant
represents a point in the future, and you want to produce a timeout that
runs until that point in time.

Changeloge: added