Lower binary expressions into regular method calls
There's no need to special-case binary expressions such as `A + B`, as
we can optimise them into native operations through inlining. By
lowering these expressions to regular calls we reduce the amount of
compiler code necessary. This also fixes a bug where `A + B` would
result in a crash if the operator was implemented such that it moves its
receiver, as the MIR pass didn't account for this.

Changelog: changed