Fix runaway recursion when inferring cyclic types
Instead of one-off hacks, we now simply limit the recursion depth when
trying to infer/resolve a type into it's final type. There are various
ways (due to the use of type placeholders and type parameters) one can
introduce cyclic types, and simply limiting the recursion depth is the
only consistent way of handling this (short of not supporting type
inference).

Changelog: fixed