Fix type-checking and inferring against traits
When comparing a class instance against a trait instance, the
implementation of the trait may contain type parameters that are
assigned values in the class instance. If we don't expose these as part
of the type-check, types might not be inferred properly.

This commit fixes this by using a sub TypeContext when comparing against
a trait instance. The TypeArguments of this context are populated
according to the type arguments of the class instance compared against
the trait instance.

Changelog: fixed