Rework iterating over byte sequences
This removes std::io::StringReader and std::io::ByteArrayReader, in
favour of String and ByteArray providing a dedicated Bytes iterator that
also implements Read. The Bytes iterator implements the new
std::iter::Bytes trait. This trait is a regular iterator, but also
provides the "next_byte" method to pull bytes out of the iterator
without wrapping them in an Option type. This makes the Bytes iterator
more efficient when dealing with large input streams.

Changelog: changed