Implement String.slice_bytes in Rust
Slicing a String using a byte range is a rather common operation, and
using a ByteArray and N calls to ByteArray.push is anything but
efficient. This commit changes the implementation to use a built-in
function written in Rust, and to return a String instead of a ByteArray.

Changelog: changed