Return nil for loop expressions in MIR
Loop expressions are typed as returning Never. This allows using them in
methods with an explicit return type, provided there's a hard return
somewhere inside the loop. For example, code such as this is valid
because of this:

    fn example -> Int {
      loop {
        return 42
      }
    }

In MIR we returned a empty register typed as Never and never wrote a
value to it. This would lead to VM crashes when using loops in an
expression context, such as inside an if. For example:

    if foo {
      loop { ... }
    } else {
      loop { ... }
    }

If we were to break out of the loop, the return value of the if/else
would be this Never register, unless the loop was followed by another
expression. This could then result in the VM using the register even
though no value was ever written to it, resulting in a crash.

To fix this, loop expressions that contain a `break` are now typed as
returning `Nil`, while loop expressions without a `break` still return
`Never`. This fixes the issue while still allowing loops to be the last
expression in a method, provided they don't contain a `break`.

Changelog: fixed