Implement cryptographic hash functions and ciphers
This commit adds the module std::crypto along with several child
modules, such as std::crypto::sha2. These modules provide support for
the following cryptographic hash functions and ciphers:

- MD5
- SHA1
- SHA256
- SHA512
- ChaCha20
- XChaCha20
- Poly1305

All implementations are written entirely in Inko, based on existing
resources such as the corresponding RFCs and the Monocypher project
(https://github.com/LoupVaillant/Monocypher/). The API overall is still
a bit rough on the edges, but for the time being it should prove
sufficient.

The implementations of ChaCha20 are based on RFC 8439 instead of the
original paper by Daniel J. Bernstein, and as such use a 96-bits nonce
instead of a 64-bits nonce.

Combinations of ChaCha20 and Poly1305 aren't provided at this time, this
is something we'll take care of in future commits.

Performance is not a goal at this stage, and as such the implementations
are likely to perform exceptionally poorly compared to other
implementations. Over time we'll optimise these implementations where
necessary.

Changelog: added