Add dedicated instruction for static method calls
Instead of using two instructions for static method calls (one to get
the receiver and one to perform a virtual method call), we now have a
dedicated CallStatic instruction. This instruction takes a class ID
directly and calls a method on the provided class.

As part of this we also remove the GetModule and GetClass instructions,
as these are no longer necessary.

Changelog: added