Document the boolean true? and false? predicates
I know the docs aren't meant to cover standard library
behaviours, but one surprising feature of Inko is that
it doesn't have a negation operator common in many languages.

This documents the `true?` and `false?` predicates with the express
purpose of mentioning how to get negation in Inko using the `false?`
predicate.