Replace the interpreter with an LLVM compiler
This replaces the bytecode interpreter with a native code compiled,
backed by LLVM. While this is a massive change, various internals remain
the same. For example, for generics we still rely on runtime pointer
checking and checks. This is a deliberate choice to keep the scope of
changes from expanding even more, and we'll take care of these remaining
issues separately.

As part of these changes we've removed the libffi based FFI, as it makes
no sense to use this API in a native code compiler. As introducing a new
FFI is going to be a lot of work, we'll also take care of this
separately, meaning you temporarily won't be able to interact with C
code.

The package manager (ipm) is also merged into the compiler executable
("inko"), removing the need for a separate command.

For more information, see https://github.com/inko-lang/inko/pull/508.