Fix Channel.try_receive not clearing its status
When Channel.try_receive was used on a channel without values, the
"waiting for channel" flag would be set but not unset, and the process
would be pushed into the waiting list. This could then result in
sporadic reschedules of the process, such as when it's simply sleeping.

This commit fixes this by using a dedicated function in the runtime for
try_receive, instead of trying to funnel all logic into a single receive
function.