Adjust reference counts when sending messages
When sending a message to a process, the reference count is first
incremented. When the message finishes, it decrements the reference
count and optionally schedules the async destructor. In addition, we
allow processes to send messages to themselves (again), as this is
necessary for processes to be able to schedule their own destruction.

This ensures we don't schedule the destruction of a process when
messages are still in flight, as those messages may result in new
references being created.

This fixes https://github.com/inko-lang/inko/issues/532 and fixes
https://github.com/inko-lang/inko/issues/533.