Fix dropping old values when assigning new ones
When assigning a new value to a field, local variable, captured local
variable or captured field, the old value is supposed to be dropped.
Prior to this commit this wasn't handled properly, and one of two
problems could arise:

1. The value simply wasn't dropped, resulting in a memory leak
2. The old value would be decremented instead of dropped, resulting in a
   reference count underflow

This commit fixes these problems by slightly tweaking how we decide if a
field or variable value needs to be dropped before assigning the new
value.

This fixes https://github.com/inko-lang/inko/issues/536.

Changelog: fixed