Verify trait requirements
The compiler didn't include a pass to verify the requirements of a
trait. This meant that something like this would be considered valid
(it's not):

    trait Foo[T] {}
    trait Bar: Foo {}

This commit fixes this by introducing a small pass that iterates over
the trait requirements, adding errors for any invalid requirements.

See https://github.com/inko-lang/inko/issues/538 for more information.

Changelog: fixed