Allow passing owned values to mutable traits
When comparing an owned value with a mutable reference to a trait, of
the owned value implements the trait we now allow passing it to a
reference to said trait. This means code such as this is now valid:

    fn example(value: mut ToSomeTrait) {}

    example(SomeThingThatImplementsToSomeTrait {})

This is valid only at the outer-most "layer" of the type, i.e. passing
`Foo` to `mut Array[ToString]` is still invalid.

The reason this is sound (or at least should be) for the outer-most type
is that the value passed can't be replaced with something of a different
type. That is, if we passed `Array[Cat]` to `mut Array[Animal]`, the
receiving method could stick a `Dog` in the `Array`. This isn't possible
for the outer-most value, so passing `Cat` to `mut Animal` is valid.

This fixes https://github.com/inko-lang/inko/issues/500.

Changelog: fixed