Make working with unique values easier
`recover` expressions now have access to values defined outside of them.
These values are exposed as `uni ref T`/`uni mut T` and come with the
same requirements as `uni T` types, such as methods only being available
of their input and output is sendable. Fields in turn are also available
for `uni T` receivers or when typed as `uni T`, and are exposed the same
way.

The result is that it's easier to work with unique values, without the
need for further complicating the type system. For example, you can now
do this:

    import std::net::ip::IpAddress

    class Config {
      let @ip: uni IpAddress
    }

    fn example(config: ref Config) -> uni IpAddress {
      recover config.ip.clone
    }

In general this should remove the need for using `uni T` as return types
all over the place, while still allowing you to easily move values
between processes.

This fixes https://github.com/inko-lang/inko/issues/528.

Changelog: added