Fix double drops for conditionally dropped values
When dropping a value conditionally, we didn't set the drop flag to
`false`, resulting in future checks potentially dropping the same value
again.

In addition, we didn't set the register state to "moved" in the block
_after_ the drop, resulting in a second check/drop being generated. We
have to set the state in the after block also, as this block will have
multiple predecessors and thus treat the register state as "maybe moved"
instead of "moved".

This fixes https://github.com/inko-lang/inko/issues/560.

Changelog: fixed