Recheck refs after processing fields
When dropping a class, we recheck the reference count to ensure a
destructor doesn't introduce new references to a value that we're
dropping. This check must be performed _after_ processing all fields and
not before it, otherwise we won't be able to drop cyclic types (e.g.
"Foo" stores an owned "Bar", which contains a reference back to "Foo").

Thanks to Dusty for reporting this in the Discord channel.

Changelog: fixed