Add Array.sort and Array.sort_by
These methods are used to sort an Array, and perform a stable sort. The
current algorithm used is the recursive merge sort algorithm. While
faster algorithms exist, they are typically much more complex to
understand. In the future we may change the algorithm, if this is deemed
worth the effort.

Changelog: added