Fix "match" not incrementing certain bindings
When matching against patterns such as `Foo(Bar(var))`, if the
outer-most variable was a reference, the pattern matching compiler
would generate match variables with an incorrect type. This would then
result in the MIR code moving a value into `var` _without_ an increment,
instead of performing an increment and moving the reference into the
variable. This would then manifest itself as a reference counting error
when dropping an owned value matched against.

This fixes https://github.com/inko-lang/inko/issues/580.

Changelog: fixed