Add support for interfacing with C libraries
This adds support for linking against C libraries and using their types
and functions. The FFI is fairly strict and a bit limited, such as not
performing automatic type conversions, but this is a deliberate choice:
it keeps the compiler's complexity at a reasonable level, and it should
(hopefully) further drive home the idea that one should avoid
interfacing with C as much as they can, as all of Inko's safety
guarantees are thrown out of the window when doing so.

In preparation for https://github.com/inko-lang/inko/issues/525, some
runtime functions return an i64 instead of an Int, while some still
return an Int. This is a little inconsistent, but the goal is to try and
reduce the amount of explicit type casts that we may need to change when
said specialisation is implemented. Once implemented, Int64 will just be
an alias for Int (or maybe I'll remove it entirely, not sure yet).

As part of this work, the precedence of type casts (`x as Type`) is
changed to be the same as binary operators. This means expressions such
as `x as Foo + y` are now valid, instead of resulting in a syntax error.
This makes working with type casts (something you'll need to use more
often when working with C code) less painful.

This commit also introduces support for conditional compilation at the
import level. Based on the build target, a set of build tags is
produced. `import` statements support a condition, and when given the
`import` is only included if all the tags match. For example:

    import foo if mac and amd64

This would only import `foo` if the tags "mac" and "amd64" are present.
OR and NOT expressions aren't supported, as one can simply use multiple
imports.

This fixes https://github.com/inko-lang/inko/issues/290.

Changelog: added