Replace use of the `libc` by the `rustix` crate
In the interest of reducing 3rd-party dependencies and maintaining
control over our code, I would not only like to restore our old polling
code in favour of the polling crate (#344), but also switch to using
rustix. The benefit of using rustix is that at least on Linux we can
avoid libc calls. While this may result in a performance improvement,
it's really more about slowly reducing our dependencies on 3rd-party
libraries, including libc. The API might also be a bit more pleasant
to use compared to using the libc crate directly.

Changelog: changed