Add proper support for FreeBSD
This adds CI configuration to test on FreeBSD, installation instructions
to handle some llvm-sys issues, and various fixes for FreeBSD to make
sure all tests pass.

As part of this we rework how the runtime handles arguments passed to
the executable. Prior to this commit, we simply used Rust's
`std::env::args_os` method, but this won't work reliably: on
Linux/glibc, Rust injects a hook to populate global data, but on FreeBSD
(and potentially other platforms) this doesn't work. To work around
this, we simply take the argc/argv arguments of the generated main()
function and pass them to the runtime.

This commit also removes any remaining traces of OpenBSD support.
Supporting OpenBSD isn't possible at this time for the following
reasons:

- We need LLVM 15, but OpenBSD only provides LLVM 13.
- To test multiple OpenBSD versions we need rustup, but rustup doesn't
  support OpenBSD (https://github.com/rust-lang/rustup/issues/2168).
- OpenBSD releases aren't compatible with each other, meaning we'd have
  to test several versions at all times. Not only does this increase the
  time to run tests, it may also make maintenance more difficult.
- I have no knowledge of using or developing for OpenBSD. While this is
  also true for FreeBSD, I found FreeBSD to at least be accessible/easy
  enough to set up.
- The overlap between OpenBSD users and those interested in Inko
  probably doesn't exist in the first place. Even if it does, the amount
  of people is probably so small it simply isn't worth the trouble.

In the future we may officially support OpenBSD, but I doubt this will
happen any time soon.

This fixes https://github.com/inko-lang/inko/issues/526.

Changelog: added