Implement Array entirely in Inko
The Array type is no longer defined in the runtime and exposed to the
standard library, instead the standard library defines the type
entirely, using the C FFI for raw memory allocations. This is a step
towards being able to implement a better way of compiling generics (see
https://github.com/inko-lang/inko/issues/525), as doing so requires
creating specialised copies of the Array type.

As part of this, array literals are desugared into regular method calls,
meaning this:

    [10, 20]

Is now turned into this:

    {
      let temp = Array.with_capacity(3)

      temp.push(10)
      temp.push(20)
      temp
    }

This removes a bunch of duplication/redundant compiler logic.

Array constants in turn are changed to use statically allocated memory
for their buffers (the actual Inko object is still heap allocated). The
permanent flag is no longer set, instead we return a pointer with the
reference bit set. While the reference count may overflow/underflow,
that's OK because we never actually check it and drop the constant. The
idea is to try and get rid of this permanent flag entirely, provided I
can find a way to not need it for string constants, as we need to
prevent dropping them when their reference count reaches zero.

Integer and float literals no longer use constants. For floats that
means heap allocations for float literals, but this is only temporary
until we implement specialisation, at which point both Int and Float
become stack/immediate values. This change makes some parts of LLVM code
generation a bit easier to work with.

The ByteArray type is still implemented in the runtime, as various Rust
functions we use specifically expect a `Vec<u8>`. Until we replace all
of that (or find another workaround), ByteArray remains defined in the
runtime library.

This fixes https://github.com/inko-lang/inko/issues/349.

Changelog: changed