Fix formatting of Int::MIN as base2/base10
Int.to_base() uses Int.absolute, but this method panics when used with
Int::MIN, as there's not positive value that can represent it. This
panic then surfaces when using Int.to_base2 or Int.to_base10 with such a
value.

To fix this, we return a known/fixed String in Int.to_base2 and
Int.to_base10 when `self` is Int::MIN.

Changelog: fixed