Implement Float.== in Inko
The implementation is the same as before, but instead of calling to a
runtime function written in Rust, we now implement this method in pure
Inko, though we still use an LLVM intrinsic for the base comparison
(exposed using the _INKO.float_eq intrinsic).

This fixes https://github.com/inko-lang/inko/issues/584.

Changelog: changed