Define more of String in Inko
This commit changes the implementation of String, both in the standard
and runtime library, such that the standard library has a bit more
knowledge/control over it. For example, the memory layout no longer
depends on a Box<[T]>, of which the memory layout isn't specified, and
instead the runtime library defines a specific layout mirrored by the
standard library. This allows removing of several runtime functions, in
favour of implementing the logic in Inko's standard library.

Changelog: changed