Replace "length" with "size"
Length is a measurement of distance, not the size or quantity. In
addition, the standard library wasn't entirely consistent in deciding
what term to use and when.

This commit changes the standard library such that "size" is always used
for quantities (or byte sizes in case of String), and "length" isn't
used at all. For String, String.characters has been renamed to
String.chars, and various "length" arguments are now called "chars".
This makes it more clear the argument refers to the number of
characters, not bytes.

Changelog: changed