Remove use of :: for namespaces
Imports and constant references no longer use "::" to separate modules
from other modules or constants, instead using ".". So instead of this:

    import foo::bar::(A, B)

You now write this:

    import foo.bar.(A, B)

And instead of this:

    errors::FOO

You now write:

    errors.FOO

The use of "." over "::" has a few benefits/reasons:

1. It's not only one character less to type, on standard keyboards it
   also removes the need for using shift
2. It's consistent with method calls and field reads
3. It removes the need for remembering "Oh wait I'm trying to access a
   constant, so I have to use ::"
4. Auto-completion engines now only need to consider "." as a completion
   start character, instead of also having to consider "::"
5. It appears less noisy compared to "::", especially when there are
   many imports grouped together

The syntax for importing symbols from a module remains the same. While I
thought about changing this, I'm not sure yet what approach to take
there, so for the time being this remains the same.

This fixes https://github.com/inko-lang/inko/issues/545.

Changelog: changed