Remove ClassId from the CallStatic instruction
This isn't needed, because when generating LLVM IR we just call the
native function directly, instead of going through a class. As part of
this, the LLVM code generation is changed a bit to not store static
methods in method tables, as they'll never be used for virtual/dynamic
dispatch.

Changelog: changed